Added unit test for sort table rows
