+++ app/client/src/components/designSystems/appsmith/TableComponent/TableUtilities.tsx
@@ -8,7 +8,6 @@ import {
 import { ColumnAction } from "components/propertyControls/ColumnActionSelectorControl";
 
 import {
-  ReactTableColumnProps,
   ColumnTypes,
   Condition,
   CellAlignmentTypes,
@@ -20,14 +19,7 @@ import {
   ConditionFunctions,
   TableStyles,
 } from "components/designSystems/appsmith/TableComponent/Constants";
-import {
-  isString,
-  isEmpty,
-  findIndex,
-  isPlainObject,
-  isNil,
-  without,
-} from "lodash";
+import { isString, isEmpty, findIndex, without } from "lodash";
 import PopoverVideo from "components/designSystems/appsmith/PopoverVideo";
 import Button from "components/editorComponents/Button";
 import AutoToolTipComponent from "components/designSystems/appsmith/TableComponent/AutoToolTipComponent";
@@ -35,7 +27,6 @@ import { ControlIcons } from "icons/ControlIcons";
 import { AnyStyledComponent } from "styled-components";
 import styled from "constants/DefaultTheme";
 import { Colors } from "constants/Colors";
-import moment from "moment";
 import { DropdownOption } from "widgets/DropdownWidget";
 import { IconNames } from "@blueprintjs/icons";
 import { Select, IItemRendererProps } from "@blueprintjs/select";
@@ -352,60 +343,6 @@ export const TableHeaderCell = (props: {
     </div>
   );
 };
-
-export function sortTableFunction(
-  filteredTableData: Array<Record<string, unknown>>,
-  columns: ReactTableColumnProps[],
-  sortedColumn: string,
-  sortOrder: boolean,
-) {
-  const tableData = filteredTableData ? [...filteredTableData] : [];
-  const columnType =
-    columns.find(
-      (column: ReactTableColumnProps) => column.accessor === sortedColumn,
-    )?.metaProperties?.type || ColumnTypes.TEXT;
-  return tableData.sort(
-    (a: { [key: string]: any }, b: { [key: string]: any }) => {
-      if (
-        isPlainObject(a) &&
-        isPlainObject(b) &&
-        !isNil(a[sortedColumn]) &&
-        !isNil(b[sortedColumn])
-      ) {
-        switch (columnType) {
-          case ColumnTypes.NUMBER:
-            return sortOrder
-              ? Number(a[sortedColumn]) > Number(b[sortedColumn])
-                ? 1
-                : -1
-              : Number(b[sortedColumn]) > Number(a[sortedColumn])
-              ? 1
-              : -1;
-          case ColumnTypes.DATE:
-            return sortOrder
-              ? moment(a[sortedColumn]).isAfter(b[sortedColumn])
-                ? 1
-                : -1
-              : moment(b[sortedColumn]).isAfter(a[sortedColumn])
-              ? 1
-              : -1;
-          default:
-            return sortOrder
-              ? a[sortedColumn].toString().toUpperCase() >
-                b[sortedColumn].toString().toUpperCase()
-                ? 1
-                : -1
-              : b[sortedColumn].toString().toUpperCase() >
-                a[sortedColumn].toString().toUpperCase()
-              ? 1
-              : -1;
-        }
-      } else {
-        return sortOrder ? 1 : 0;
-      }
-    },
-  );
-}
 
 export function compare(a: any, b: any, condition: Condition) {
   let result = true;
